% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyQuickStarter.R
\name{shinyQuickStarter}
\alias{shinyQuickStarter}
\title{Starts the ShinyQuickStarter Addin in RStudio.}
\usage{
shinyQuickStarter()
}
\description{
Starts the ShinyQuickStarter Addin in RStudio.
}
\examples{
\dontrun{shinyQuickStarter()}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}

Bernhard Daffner \email{bernhard.daffner@th-deg.de}

Joerg Bauer \email{joerg.bauer@th-deg.de}
}
