% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyQuickStarter-package.R
\docType{package}
\name{ShinyQuickStarter-package}
\alias{ShinyQuickStarter-package}
\alias{ShinyQuickStarter}
\title{'RStudio' Addin for Building Shiny Apps per Drag & Drop}
\description{
This 'RStudio' addin makes the creation of 'Shiny' and 'ShinyDashboard' apps more efficient.  Besides the necessary folder structure, entire apps can be created using a drag and drop interface and customized with respect to a specific use case. The addin allows the export of the required user interface and server code at any time. By allowing the creation of modules, the addin can be used throughout the entire app development process.
}
\details{
\tabular{ll}{ Package: \tab ShinyQuickStarter\cr Type: \tab Package\cr Version:
\tab 2.0.0\cr Date: \tab 2020-03-02\cr License: \tab GPL-3\cr Depends: \tab
R (>= 3.6.0)}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}

Bernhard Daffner \email{bernhard.daffner@th-deg.de}

Joerg Bauer \email{joerg.bauer@th-deg.de}
}
